requirements.txt
pandas>=1.5.3
pygame>=2.5.0
pyttsx3>=2.90
pyserial>=3.5
mysql-connector-python>=8.0.0 